# backend/liive-user-manager/docs/swagger.yaml
basePath: /
definitions:
  types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  types.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  types.UpdatePasswordRequest:
    properties:
      currentPassword:
        type: string
      newPassword:
        minLength: 8
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
  types.UpdateProfileRequest:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    type: object
  types.UserResponse:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      lastName:
        type: string
      username:
        type: string
    type: object
host: localhost:8083
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the user management service API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User Manager API
  version: "1.0"
paths:
  /api/password:
    put:
      consumes:
      - application/json
      description: Update user password with current and new password
      parameters:
      - description: Password update details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user password
      tags:
      - users
  /api/profile:
    put:
      consumes:
      - application/json
      description: Update user profile information
      parameters:
      - description: User profile update details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, username and password
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Register a new user
      tags:
      - users
schemes:
- http
swagger: "2.0"